# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Allbtc Swagger UI",
        "description": "Документация OpenApi для Allbtc pool",
        "contact": {
            "email": "ognevoy.bima@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/v1/"
        }
    ],
    "paths": {
...
        "/watchers/{user}/{sub}": {
            "get": {
                "tags": [
                    "Watcher Links"
                ],
                "summary": "Get watcher links for a user's sub",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User's ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sub",
                        "in": "path",
                        "description": "Sub's ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WatcherLinkResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "User or sub not found"
                    }
                }
            }
        },
        "/workers/{sub}": {
            "get": {
                "tags": [
                    "Worker"
                ],
                "summary": "Get sub's worker list",
                "parameters": [
                    {
                        "name": "sub",
                        "in": "path",
                        "description": "Sub's ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter workers by status (all, active, inactive)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "all",
                                "active",
                                "inactive"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "gid": {
                                                "type": "integer"
                                            },
                                            "group_name": {
                                                "type": "string"
                                            },
                                            "worker_id": {
                                                "type": "string"
                                            },
                                            "worker_name": {
                                                "type": "string"
                                            },
                                            "is_top": {
                                                "type": "boolean"
                                            },
                                            "shares_1m": {
                                                "type": "string"
                                            },
                                            "shares_5m": {
                                                "type": "string"
                                            },
                                            "shares_15m": {
                                                "type": "string"
                                            },
                                            "accept_count": {
                                                "type": "string"
                                            },
                                            "last_share_time": {
                                                "type": "integer"
                                            },
                                            "last_share_ip": {
                                                "type": "string"
                                            },
                                            "reject_percent": {
                                                "type": "string"
                                            },
                                            "first_share_time": {
                                                "type": "integer"
                                            },
                                            "miner_agent": {
                                                "type": "string"
                                            },
                                            "shares_unit": {
                                                "type": "string"
                                            },
                                            "status": {
                                                "type": "string"
                                            },
                                            "shares_1h": {
                                                "type": "integer"
                                            },
                                            "shares_1d": {
                                                "type": "string"
                                            },
                                            "shares_1m_pure": {
                                                "type": "string"
                                            },
                                            "shares_5m_pure": {
                                                "type": "string"
                                            },
                                            "shares_15m_pure": {
                                                "type": "string"
                                            },
                                            "shares_1h_pure": {
                                                "type": "integer"
                                            },
                                            "shares_1d_pure": {
                                                "type": "integer"
                                            },
                                            "shares_1d_unit": {
                                                "type": "string"
                                            },
                                            "reject_percent_1d": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Sub not found"
                    }
                }
            }
        },
        "/worker/{worker}": {
            "get": {
                "tags": [
                    "Worker"
                ],
                "summary": "Get worker details",
                "parameters": [
                    {
                        "name": "worker",
                        "in": "path",
                        "description": "Worker's ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "worker_id": {
                                                    "type": "string"
                                                },
                                                "worker_name": {
                                                    "type": "string"
                                                },
                                                "shares_1m": {
                                                    "type": "string"
                                                },
                                                "shares_5m": {
                                                    "type": "string"
                                                },
                                                "shares_15m": {
                                                    "type": "string"
                                                },
                                                "accept_count": {
                                                    "type": "integer"
                                                },
                                                "accept_percent": {
                                                    "type": "integer"
                                                },
                                                "reject_percent": {
                                                    "type": "string"
                                                },
                                                "last_share_ip": {
                                                    "type": "string"
                                                },
                                                "ip2location": {
                                                    "type": "string"
                                                },
                                                "last_share_time": {
                                                    "type": "integer"
                                                },
                                                "shares_unit": {
                                                    "type": "string"
                                                },
                                                "worker_status": {
                                                    "type": "string"
                                                },
                                                "shares_1h": {
                                                    "type": "integer"
                                                },
                                                "shares_1d": {
                                                    "type": "string"
                                                },
                                                "shares_1m_pure": {
                                                    "type": "string"
                                                },
                                                "shares_5m_pure": {
                                                    "type": "string"
                                                },
                                                "shares_15m_pure": {
                                                    "type": "string"
                                                },
                                                "shares_1h_pure": {
                                                    "type": "integer"
                                                },
                                                "shares_1d_pure": {
                                                    "type": "integer"
                                                },
                                                "shares_1d_unit": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Worker not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AllowedRouteResource": {
                "properties": {
                    "allowed_routes": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
...
            "WatcherLinkResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "allowed_routes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AllowedRouteResource"
                        }
                    },
                    "access_count": {
                        "type": "integer"
                    },
                    "url": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}